# [Optimize] reduce pilot memory in large degree