@istio-testing 2021-12-07
@hzxuzhonghu: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
lint_istio | 325e9ead123e859ed5a05305dff3620e354eac9d | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/36400/lint_istio/1468061779971018752) | true | `/test lint_istio`
release-notes_istio | eb586847f44403361d82c463aebe73884c0ff750 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/36400/release-notes_istio/1468133334486880256) | true | `/test release-notes_istio`
unit-tests_istio | eb586847f44403361d82c463aebe73884c0ff750 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/36400/unit-tests_istio/1468133325259411456) | true | `/test unit-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->