@howardjohn 2021-12-07
I am a bit confused by the performance numbers. How can this impact pilot_xds_push_time_bucket? we are still running the same push code...? Just less GC pressure?

Also confused how stable memory usage is reduced. If its stable there should be no requests so where is the memory differences?