@hzxuzhonghu 2021-12-08
>How can this impact pilot_xds_push_time_bucket? 
 i guess it has something todo with the client receiver, 

```
	// SendMsg blocks until:
	//   - There is sufficient flow control to schedule m with the transport, or
	//   - The stream is done, or
	//   - The stream breaks.
```


>Also confused how stable memory usage is reduced. If its stable there should be no requests so where is the memory differences?

I am not sure either, it is a bonne surprise