@hzxuzhonghu 2021-12-16
With 1.12 the result is different,  

|  | Before | After |
| ------ | ------ | ------ |
| Peek | 13.96GB|  11.7 GB |
| Stable | 7.56 GiB | 7.36 GiB |